breakpoints almost work

+++ plugins-client/ext.debugger/breakpoints.js
@@ -33,6 +33,7 @@ module.exports = {
             }
             // bind it to the Breakpoint model
             mdlDbgBreakpoints.load(bpFromIde);
+            _self.$syncOpenFiles();
         });
 
         // register dock panel
@@ -49,20 +50,15 @@ module.exports = {
             return dbgBreakpoints;
         });
         
-        // ide.addEventListener("afteropenfile", evHandler);
         ide.addEventListener("afterfilesave", function(e) {
             var page = e.nextPage;
             if (!page || !page.$editor || !page.$editor.ceEditor)
                 return;
-            var ace = page.$editor.ceEditor.$editor
-            if (!ace.$breakpointListener)
-                _self.initEditor(ace);
-            
-            if (!ace.session.$breakpointListener)
-                _self.initSession(ace.session);
-
-            _self.updateSession(ace.session, page.$doc.getNode());
+            var ace = page.$editor.ceEditor.$editor         
+            if (ace.session.$breakpoints.length)
+                _self.updateBreakpointModel(session);
         });
+        
         ide.addEventListener("tab.afterswitch", function(e) {
             var page = e.nextPage;
             if (!page || !page.$editor || !page.$editor.ceEditor)
@@ -102,9 +98,10 @@ module.exports = {
     
     initEditor: function(editor) {
         var _self = this;
-        var el = document.createElement("div");
-        editor.renderer.$gutter.appendChild(el);
-        el.style.cssText = "position:absolute;top:0;bottom:0;left:0;width:18px;cursor:pointer"
+        //todo this breaks annotation tooltip
+        //var el = document.createElement("div");
+        //editor.renderer.$gutter.appendChild(el);
+        //el.style.cssText = "position:absolute;top:0;bottom:0;left:0;width:18px;cursor:pointer"
         
         editor.on("guttermousedown", editor.$breakpointListener = function(e) {
             if (!editor.isFocused())
@@ -116,17 +113,13 @@ module.exports = {
 
             var session = editor.session;
             var bp = session.getBreakpoints()[row];
-            if (!bp)
-                bp = " ace_breakpoint ";
-            else if(bp.indexOf("disabled") == -1)
-                bp = " ace_breakpoint disabled ";
-            else
-                bp = null;
+            var i = bp ? bp.indexOf("disabled") == -1 ? 1 : 2 : 0;
+            if (e.getShiftKey())
+                i = (i + 1) %3;
+            bp = [" ace_breakpoint ", " ace_breakpoint disabled ", null][i];
 
             session.setBreakpoint(row, bp);
-            
-            session.getBreakpoints();
-            session.c9doc.getNode();
+            _self.updateBreakpointModel(session);
         });
     },
     initSession: function(session) {
@@ -147,7 +140,7 @@ module.exports = {
 
 			if (len > 0) {
 				args = Array(len);
-				args.unshift(firstRow, 0)
+				args.unshift(firstRow, 0);
 				this.$breakpoints.splice.apply(this.$breakpoints, args);
             } else if (len < 0) {
                 var rem = this.$breakpoints.splice(firstRow + 1, -len);
@@ -167,15 +160,14 @@ module.exports = {
         var rows = [];
         if (node) {
             var path = node.getAttribute("path");
-            var scriptPath = path.slice(ide.davPrefix.length);
-            var breakpoints = mdlDbgBreakpoints.queryNodes("//breakpoint[@scriptPath='" + scriptPath + "']");
+            var breakpoints = mdlDbgBreakpoints.queryNodes("//breakpoint[@path='" + path + "']");
 
             for (var i=0; i< breakpoints.length; i++) {
                 var bp = breakpoints[i]
                 var line = parseInt(bp.getAttribute("line"), 10);
                 var offset = parseInt(bp.getAttribute("lineoffset"), 10);
                 var enabled = apf.isTrue(bp.getAttribute("enabled"));
-                rows[line] = " ace_breakpoint " + enabled ? "" : "disabled ";
+                rows[line] = " ace_breakpoint " + (enabled ? "" : "disabled ");
             }
         }
         session.setBreakpoints(rows);
@@ -192,41 +184,71 @@ module.exports = {
         
     },
     
-    toggleBreakpoint : function(script, row, content) {
-        var scriptName = script.getAttribute("scriptname");
-        var bp = model.queryNode("breakpoint[@script='" + scriptName
+    removeBreakpoint: function(path, row) {
+        var bp = mdlDbgBreakpoints.queryNode("breakpoint[@path='" + path
             + "' and @line='" + row + "']");
-
-        if (bp) {
-            apf.xmldb.removeNode(bp);
-        }
-        else {
-            // filename is something like blah/blah/workspace/realdir/file
-            // we are only interested in the part after workspace for display purposes
-            var tofind = "/workspace/";
-            var path = script.getAttribute("path");
-            var displayText = path;
-            if (path.indexOf(tofind) > -1) {
-                displayText = path.substring(path.indexOf(tofind) + tofind.length);
-            }
-
-            var bp = apf.n("<breakpoint/>")
-                .attr("script", scriptName)
-                .attr("line", row)
-                .attr("text", displayText + ":" + (parseInt(row, 10) + 1))
-                .attr("lineoffset", 0)
-                .attr("content", content)
-                .attr("enabled", "true")
-                .node();
-            model.appendXml(bp);
-        }
+        bp && apf.xmldb.removeNode(bp);
     },
+    
+    addBreakpoint: function(path, row, content) {
+        var displayText = path;
+        var tofind = ide.davPrefix;
+        if (path.indexOf(tofind) > -1) {
+            displayText = path.substring(path.indexOf(tofind) + tofind.length);
+        }
 
+        var bp = apf.n("<breakpoint/>")
+            .attr("path", path)
+            .attr("line", row)
+            .attr("text", displayText + ":" + (row + 1))
+            .attr("lineoffset", 0)
+            .attr("content", content)
+            .attr("enabled", "true")
+            .node();
+        mdlDbgBreakpoints.appendXml(bp);
+    },
+    
+    removeBreakpoint: function(path, row) {
+        var bp = mdlDbgBreakpoints.queryNode("breakpoint[@path='" + path
+            + "' and @line='" + row + "']");
+        bp && apf.xmldb.removeNode(bp);
+    },
+    
+    updateBreakpointModel: function(session) {
+        var path = session.c9doc.getNode().getAttribute("path");
+        var breakpoints = session.$breakpoints;
+        var displayText = path;
+        var tofind = ide.davPrefix;
+        if (path.indexOf(tofind) == 0)
+            displayText = path.substring(tofind.length + 1);
+        
+        var bpList = mdlDbgBreakpoints.queryNodes("breakpoint[@path='" + path + "']");
+        for (var i = bpList.length; i--; ) {
+            apf.xmldb.removeNode(bpList[i]);
+        }
+        // iterate over sparse array
+        breakpoints.forEach(function(breakpoint, row) {
+            if (!breakpoint)
+                return
+            var bp = apf.n("<breakpoint/>")
+                .attr("path", path)
+                .attr("line", row)
+                .attr("text", displayText + ":" + (+row + 1))
+                .attr("lineoffset", 0)
+                .attr("content", session.getLine(row))
+                .attr("enabled", breakpoint.indexOf("disabled") == -1)
+                .node();
+            mdlDbgBreakpoints.appendXml(bp);
+        })
+    },
+    
     setBreakPointEnabled : function(node, value){
         node.setAttribute("enabled", value ? true : false);
     },
     
-    $syncTree: function() {
+    $syncOpenFiles: function() {
+        
+        return
         if (this.inSync) return;
         this.inSync = true;
         var dbgFiles = mdlDbgSources.data.childNodes;
@@ -253,7 +275,7 @@ module.exports = {
                 apf.b(file).attr("scriptid", dbgFile.getAttribute("scriptid"));
         }
         this.inSync = false;
-    },
+    }
 }
 
 });
